% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{logbin}
\alias{logbin}
\title{logbin}
\usage{
logbin(x, param)
}
\arguments{
\item{x}{x values for the function}

\item{param}{parameter values to be sampled}
}
\value{
a vector
}
\description{
log likelihood function for binomials
}
\examples{
\dontrun{logbin(x = c(3,4,5), prob = param)}
}
