% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{logexp}
\alias{logexp}
\title{logexp}
\usage{
logexp(x, param)
}
\arguments{
\item{x}{x values for the function}

\item{param}{parameter values to be sampled}
}
\value{
a vector
}
\description{
log likelihood function for exponentials
}
\examples{
\dontrun{logexp(x = c(3,4,5), rate = param)}
}
