% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{logpoiss}
\alias{logpoiss}
\title{logpoiss}
\usage{
logpoiss(x, param)
}
\arguments{
\item{x}{x values for the function}

\item{param}{parameter values to be sampled}
}
\value{
a vector
}
\description{
log likelihood function for Poisson distributions
}
\examples{
\dontrun{logpoiss(x = c(3,4,5), lambda = param)}
}
