% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{mymaxlikg}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{function to be used}

\item{theta}{parameter to be estimated}
}
\value{
a plot and a list
}
\description{
a more general function for a grid approximation to make maximum likelihood estimations, provided you supply the appropriate helper function
}
\examples{
\dontrun{mymaxlikg(theta=seq(0,1,length=10000))}
}
