% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymld.R
\name{mymld}
\alias{mymld}
\title{mymld}
\usage{
mymld(lfun, x, param, delta = 1e-04)
}
\arguments{
\item{lfun}{function to be used}

\item{x}{x values for the function}

\item{param}{parameter values to be sampled}

\item{delta}{accuracy level}
}
\value{
a plot and a list
}
\description{
an improved function that will increase accuracy of approximations to the `mymaxlik` function
}
\examples{
\dontrun{mymld(x = c(5,5,6,7),param=seq(0,100,length=1000),lfun=logpoiss)}
}
