% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{x values for the function}

\item{mu}{sampled values for the mean}

\item{sig}{sampled values for the sd}

\item{...}{pass additional options to the function}
}
\value{
a plot and a list
}
\description{
a function which will enable you to find the maximum likelihood estimates for mu and sigma when the data are from a Normal distribution
}
\examples{
\dontrun{mymlnorm(x=c(5),mu=seq(5,10,length=100),sig=seq(.1,4,length=100),lwd=2,labcex=1)}
}
