% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNEWnewt.R
\name{myNEWnewt}
\alias{myNEWnewt}
\title{myNEWnewt}
\usage{
myNEWnewt(x0, delta = 0.001, f, xlim = c(-x0, x0))
}
\arguments{
\item{x0}{initial value}

\item{delta}{accuracy}

\item{f}{function}

\item{xlim}{limitations on x}
}
\value{
a plot and a list
}
\description{
my take on `mynewt`, a function to find the derivative of f without having to input `fdash`
}
\examples{
\dontrun{myNEWnewt(x0 = 10, delta = 0.001, f = function(x) x^2 - 5*x + 6)}
}
