% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynewt.R
\name{mynewt}
\alias{mynewt}
\title{mynewt}
\usage{
mynewt(x0, delta = 0.001, f, fdash, xlim = c(-x0, x0))
}
\arguments{
\item{x0}{initial value}

\item{delta}{accuracy}

\item{f}{function}

\item{fdash}{derivative of f}

\item{xlim}{limitations on x}
}
\value{
a plot and a list
}
\description{
a function which will find roots of a function as long as you are careful to start the algorithm with an appropriate initial value and have the derivative
}
\examples{
\dontrun{mynewt(x0 = 10,delta = 0.000001,f = function(x) x^2-4,fdash = function(x) 2*x )}
}
