% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{myNRML}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{initial value}

\item{delta}{accuracy}

\item{llik}{log likelihood function}

\item{xrange}{range for x}

\item{parameter}{parameter to be sampled}
}
\value{
a plot and a list
}
\description{
another take on the `mynewt` function, creates derivatives internally, given to us
}
\examples{
\dontrun{myNRML(x0=0.99,delta=0.000001,llik="mean",xrange=c(0.01,0.99),parameter="p")}
}
