% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypointest.R
\name{shinypointest}
\alias{shinypointest}
\title{shinypointest}
\usage{
shinypointest()
}
\value{
a plot
}
\description{
An interactive shiny app which allows control over the alpha level and iterations
}
\examples{
\dontrun{shinypointest()}
}
